requirements.txt
Flask
mysql-connector-python
gunicorn
numpy
pandas
scikit-learn
